requirements.txt
python-dotenv>=1.0.0
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2
matplotlib>=3.6
seaborn>=0.12
streamlit>=1.20
pyyaml>=6.0
gunicorn>=20.1
python-pptx>=0.6.21
openpyxl>=3.0
pytest>=7.0
black>=23.1
